fix: validations
